Finish API

+++ API_TOPOLOGY/M1.cs
@@ -10,8 +10,8 @@ namespace API_TOPOLOGY
     class M1
     {
         [JsonProperty(PropertyName = "default")]
-        public int def { get; set; }
-        public int min { get; set; }
-        public int max { get; set; }
+        public double def { get; set; }
+        public double min { get; set; }
+        public double max { get; set; }
     }
 }
